- Switched Android toolchains stl to libc++.

+++ build_android.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
 # User configuration
-ANDROID_NDK_ROOT="${ANDROID_HOME}/ndk-bundle"
 ANDROID_API="android-21"
 
 # Check settings
@@ -40,7 +39,7 @@ ARCH_NAME=("arm" "armv7" "arm64" "mips" "mips64" "x86" "x86_64")
 # Prepare toolchains
 for ((i=0; i<${#TOOLCHAIN_ARCH[@]}; i++)); do
 	if [ ! -d ${TOOLCHAIN_DIR}/${TOOLCHAIN_ARCH[$i]} ]; then
-		${ANDROID_NDK_ROOT}/build/tools/make-standalone-toolchain.sh --arch=${TOOLCHAIN_ARCH[$i]} --platform=${ANDROID_API} --install-dir=${TOOLCHAIN_DIR}/${TOOLCHAIN_ARCH[$i]}
+		${ANDROID_NDK_ROOT}/build/tools/make-standalone-toolchain.sh --arch=${TOOLCHAIN_ARCH[$i]} --platform=${ANDROID_API} --stl=libc++ --install-dir=${TOOLCHAIN_DIR}/${TOOLCHAIN_ARCH[$i]}
 	fi
 done
 
@@ -59,8 +58,13 @@ for ((i=0; i<${#ARCH[@]}; i++)); do
     if [ "${ARCH[$i]}" = "linux64-mips64" ]; then
 		export CXXFLAGS="${CXXFLAGS} -fintegrated-as"
 	fi
+	
+	if [ ! -d ${WORKING_DIR}/lib/android/${ARCH_NAME[$i]} ]; then
+		mkdir -p ${WORKING_DIR}/../../lib/android/${ARCH_NAME[$i]}
+	else
+		rm -f ${WORKING_DIR}/lib/android/${ARCH_NAME[$i]}/libhermes.a
+	fi
 
-	rm -f ${WORKING_DIR}/lib/android/${ARCH_NAME[$i]}/libhermes.a
 	rm -rf ${TMP_DIR}
     mkdir ${TMP_DIR}
