+++ depend/openssl/build_android.sh
@@ -2,7 +2,6 @@
 
 # User configuration
 LIBRARY_NAME="openssl-1.1.0f"
-ANDROID_NDK_ROOT="${ANDROID_HOME}/ndk-bundle"
 ANDROID_API="android-21"
 
 # Check settings
@@ -46,7 +45,7 @@ ARCH_NAME=("arm" "armv7" "arm64" "mips" "mips64" "x86" "x86_64")
 # Prepare toolchains
 for ((i=0; i<${#TOOLCHAIN_ARCH[@]}; i++)); do
 	if [ ! -d ${TOOLCHAIN_DIR}/${TOOLCHAIN_ARCH[$i]} ]; then
-		${ANDROID_NDK_ROOT}/build/tools/make-standalone-toolchain.sh --arch=${TOOLCHAIN_ARCH[$i]} --platform=${ANDROID_API} --install-dir=${TOOLCHAIN_DIR}/${TOOLCHAIN_ARCH[$i]}
+		${ANDROID_NDK_ROOT}/build/tools/make-standalone-toolchain.sh --arch=${TOOLCHAIN_ARCH[$i]} --platform=${ANDROID_API} --stl=libc++ --install-dir=${TOOLCHAIN_DIR}/${TOOLCHAIN_ARCH[$i]}
 	fi
 done
 
@@ -77,8 +76,13 @@ for ((i=0; i<${#ARCH[@]}; i++)); do
 	export CFLAGS="${ARCH_FLAGS} -O2 -fPIC -fno-strict-aliasing -fstack-protector"
 	export LDFLAGS="${ARCH_LINK}"
 
-	rm -f ${WORKING_DIR}/../../lib/android/${ARCH_NAME[$i]}/libcrypto.a
-	rm -f ${WORKING_DIR}/../../lib/android/${ARCH_NAME[$i]}/libssl.a
+	if [ ! -d ${WORKING_DIR}/lib/android/${ARCH_NAME[$i]} ]; then
+		mkdir -p ${WORKING_DIR}/../../lib/android/${ARCH_NAME[$i]}
+	else
+		rm -f ${WORKING_DIR}/../../lib/android/${ARCH_NAME[$i]}/libcrypto.a
+		rm -f ${WORKING_DIR}/../../lib/android/${ARCH_NAME[$i]}/libssl.a
+	fi
+	
 	rm -rf ${TMP_DIR}
 
 	./Configure ${ARCH[$i]} --prefix=${TMP_DIR} \
